Synchronized entire operations to GPS seconds PPS
This was a lot of work to get right, but it now starts the loop exactly every two seconds synchronized on the GPS Pulse-Per-Second signal line.  This replaces the unreliable delay(1000) at the end of the loop.  It does not handle the GPS missing well, it should fall back to the millis counter, but will take care of that in the future.  It's currently capturing GGA only