GPS logic rearranged and string rearranged, output working
Made the GPS work more like Adafruit background, less complpicated, reading in one char per main loop, non-blocking.  Previously was a blocking read waiting for PPS or millis.  Now it uses library calls to get the date, time, lat/lon alt, fix quality and put it into the write_to_file_string.  This is not yet evaluated for dealing with GPS errors/missing etc.  It works when the GPS is working, that's all that's been tried.