Added boot/run LED code, pinmodes cleanup
LED comes on at the beginning of the Setup loop and turns off at the end.  It flashes during normal operation; it turns on when sensor reading starts and turns off after data is written and output.  Also moved the pinmodes around and made a #define for GPS_PPS_PIN